requirements.txt
python-telegram-bot==21.4
openai>=1.40.0
python-dotenv>=1.0.1
httpx>=0.27.0
anyio>=4.4.0
# Optional MCP server scaffold
mcp>=1.2.0
# Bybit integration is left as a stub. Add a client as needed.
numpy>=1.26.0
pandas>=2.2.2
xgboost>=2.1.1
